requirements.txt
Django==3.2.12
djangorestframework==3.12.4
Pillow==8.4.0
google-generativeai==0.3.2
